Refactor: segregate dto to extra package
