+++ src/main/java/techcourse/myblog/web/dto/UserRequestDto.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.web;
+package techcourse.myblog.web.dto;
 
 public class UserRequestDto {
     private String name;
