+++ src/main/java/techcourse/myblog/web/dto/UserUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.web;
+package techcourse.myblog.web.dto;
 
 public class UserUpdateRequestDto {
     private String name;
