+++ src/test/java/techcourse/myblog/web/UserControllerTest.java
@@ -8,6 +8,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.reactive.server.EntityExchangeResult;
 import org.springframework.test.web.reactive.server.WebTestClient;
 import org.springframework.web.reactive.function.BodyInserters;
+import techcourse.myblog.web.dto.UserRequestDto;
 
 import java.util.function.Consumer;
